refactor(executor): move inline unit tests to src/tests/ module
Relocates the three inline `#[cfg(test)] mod tests` blocks in executor/src
into a dedicated `executor/src/tests/` module, matching the convention
already used by the prover and math crates.

- flamegraph.rs                -> tests/flamegraph_tests.rs (demangle)
- vm/instruction/execution.rs  -> tests/keccak_tests.rs (keccak-f[1600] + syscall)
- vm/memory.rs                 -> tests/memory_tests.rs (public output, load bounds)

`demangle` and `Memory::set_bytes_aligned` were module-private; bumped to
pub(crate) so the relocated tests can reach them. No test logic changed.
