+++ executor/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod elf;
 pub mod flamegraph;
+pub mod tests;
 pub mod vm;
