+++ executor/src/tests/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(test)]
+pub mod flamegraph_tests;
+#[cfg(test)]
+pub mod keccak_tests;
+#[cfg(test)]
+pub mod memory_tests;
